% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdaStableNet-package.R
\docType{package}
\name{AdaStableNet-package}
\alias{AdaStableNet-package}
\title{AdaStableNet: Stable Estimation of Linear ODE Systems}
\description{
AdaStableNet estimates a constant matrix \code{A} in the autonomous linear system
\verb{dX(t) / dt = A X(t)} from a noisy multivariate trajectory. It combines a
functional-data initializer with a profiled real-modal nonlinear
least-squares fit. See \code{\link[=FitAdaStableNet]{FitAdaStableNet()}} for the main interface,
\code{\link[=predict.adastablenet_fit]{predict.adastablenet_fit()}} for forecasting, and
\code{\link[=simulate_adastablenet]{simulate_adastablenet()}} for reproducible experiments.
}
\references{
Package methodology and citation information are available through
\code{citation("AdaStableNet")}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LopezRicardo1/AdaStableNet}
  \item Report bugs at \url{https://github.com/LopezRicardo1/AdaStableNet/issues}
}

}
\author{
\strong{Maintainer}: Ricardo J. Lopez Candelaria \email{rlopez.research@gmail.com}

}
\keyword{internal}
