% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.adastablenet_fit}
\alias{summary.adastablenet_fit}
\title{Summarize an AdaStableNet Fit}
\usage{
\method{summary}{adastablenet_fit}(object, ...)
}
\arguments{
\item{object}{An \code{adastablenet_fit} object.}

\item{...}{Unused.}
}
\value{
An object of class \code{summary.adastablenet_fit} with a branch table.
}
\description{
Summarize an AdaStableNet Fit
}
