% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeinv.R
\name{ridge.inv}
\alias{ridge.inv}
\title{Ridge-Inverse of a Symmetric Matrix}
\usage{
ridge.inv(SymMat, lambda.prop = 0.1^4)
}
\arguments{
\item{SymMat}{A symmetric square matrix}

\item{lambda.prop}{Proportion of trace used as ridge penalty (default = 0.0001)}
}
\value{
A regularized inverse of SymMat
}
\description{
Computes the stabilized inverse of a symmetric matrix using ridge regularization.
Used in PCODE and AdaStableNet for basis inversion.
}
